Add files via upload
The "TestMatrix.txt" refers to the record name and corresponding flow regime annotation. There are two columns in the "TestMatrix.txt" file. The first column refers to the file name in the "txt.zip", and the second column refers to the corresponding annotations. For example, the first row, "5Air0.1Water.txt	1", indicates that the record file name is "5Air0.1Water.txt", and the flow regime annotation is "1".